@@ -7,7 +7,7 @@ import 'package:weather/src/models/index.dart';
 import 'package:weather/src/presentation/search_location_page.dart';
 
 class HomePage extends StatelessWidget {
-  const HomePage({Key key}) : super(key: key);
+  const HomePage({Key? key}) : super(key: key);
 
   String _getTemperature(bool isMetricSystem, double temperature) {
     if (isMetricSystem) {
@@ -47,32 +47,43 @@ class HomePage extends StatelessWidget {
                 children: <Widget>[
                   const Spacer(),
                   Text(
-                    weatherViewModel.locationForecast.title,
+                    weatherViewModel.locationForecast!.title,
                     style: const TextStyle(fontSize: 24),
                   ),
-                  Text(_getTemperature(weatherViewModel.isMetricSystem, weatherViewModel.locationForecast.temperature)),
-                  Text('${weatherViewModel.locationForecast.weather}'),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      Text(_getTemperature(
+                          weatherViewModel.isMetricSystem, weatherViewModel.locationForecast!.temperature)),
+                      Text(weatherViewModel.isMetricSystem ? ' °C' : ' °F'),
+                    ],
+                  ),
+                  Text('${weatherViewModel.locationForecast!.weather}'),
                   IconButton(
                     icon: const Icon(Icons.refresh),
                     onPressed: () {
                       StoreProvider.of<AppState>(context)
-                          .dispatch(GetForecast(weatherViewModel.locationForecast.woeid));
+                          .dispatch(GetForecast(weatherViewModel.locationForecast!.woeid));
                     },
                   ),
                   const Spacer(),
                   Row(
                     children: <Widget>[
                       const Spacer(),
-                      RaisedButton(
-                        color: !weatherViewModel.isMetricSystem ? Colors.blue : Colors.grey,
+                      ElevatedButton(
+                        style: ElevatedButton.styleFrom(
+                          primary: !weatherViewModel.isMetricSystem ? Colors.blue : Colors.grey,
+                        ),
                         onPressed: () {
                           StoreProvider.of<AppState>(context).dispatch(const SetIsMetricSystem(false));
                         },
                         child: const Text('°F'),
                       ),
                       const Spacer(),
-                      RaisedButton(
-                        color: weatherViewModel.isMetricSystem ? Colors.blue : Colors.grey,
+                      ElevatedButton(
+                        style: ElevatedButton.styleFrom(
+                          primary: weatherViewModel.isMetricSystem ? Colors.blue : Colors.grey,
+                        ),
                         onPressed: () {
                           StoreProvider.of<AppState>(context).dispatch(const SetIsMetricSystem(true));
                         },
