@@ -1,11 +1,11 @@
 part of actions;
 
 @freezed
-abstract class GetForecast with _$GetForecast implements AppAction {
+class GetForecast with _$GetForecast implements AppAction {
   const factory GetForecast(int woeid) = GetForecast$;
 
   const factory GetForecast.successful(LocationForecast locationForecast) = GetForecastSuccessful;
 
   @Implements(ErrorAction)
-  const factory GetForecast.error(Object error) = GetForecastError;
+  const factory GetForecast.error(Object error, StackTrace stackTrace) = GetForecastError;
 }
