@@ -1,6 +1,6 @@
 part of actions;
 
 @freezed
-abstract class SetIsMetricSystem with _$SetIsMetricSystem {
+ class SetIsMetricSystem with _$SetIsMetricSystem {
   const factory SetIsMetricSystem(bool isMetric) = _SetIsMetricSystem;
 }
