@@ -3,8 +3,7 @@ library containers;
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
-import 'package:meta/meta.dart';
-import 'package:weather/src/models/index.dart';
 import 'package:redux/redux.dart';
+import 'package:weather/src/models/index.dart';
 
 part 'weather_view_model_container.dart';
