@@ -11,7 +11,8 @@ Serializers _$serializers = (new Serializers().toBuilder()
       ..add(Location.serializer)
       ..add(LocationForecast.serializer)
       ..addBuilderFactory(
-          const FullType(BuiltList, const [const FullType(Location)]), () => new ListBuilder<Location>()))
+          const FullType(BuiltList, const [const FullType(Location)]),
+          () => new ListBuilder<Location>()))
     .build();
 
-// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,lines_longer_than_80_chars,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new
+// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,deprecated_member_use_from_same_package,lines_longer_than_80_chars,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new
